+++ src/core/plugins/core-commands/server/index.ts
@@ -4,6 +4,7 @@ import { ConsoleCommander } from '@AthenaShared/utility/consoleCommander';
 
 import './commands/moderator';
 import './commands/player';
+import './commands/admin';
 
 const PLUGIN_NAME = 'commands';
 Athena.systems.plugins.registerPlugin(PLUGIN_NAME, () => {